+++ tests/test_resources.py
@@ -108,12 +108,11 @@ class TestResources(unittest.TestCase):
         self.response.json.return_value = responses['attachment']['get']
         self.assertEqual(self.redmine.attachment.get(1).export_url('pdf'), None)
 
+    @mock.patch('redmine.open', mock.mock_open(), create=True)
     def test_export(self):
-        self.response.content = 'foo'
-        self.response.text = b'foo'.decode()
         self.response.json.return_value = responses['issue']['get']
-        self.assertEqual(self.redmine.issue.get(1).export('txt', 'bytes'), 'foo')
-        self.assertEqual(self.redmine.issue.get(1).export('txt', 'text'), b'foo'.decode())
+        self.response.iter_content = lambda chunk_size: (str(num) for num in range(0, 5))
+        self.assertEqual(self.redmine.issue.get(1).export('txt', '/foo/bar'), '/foo/bar/1.txt')
 
     def test_export_not_supported_exception(self):
         from redmine.exceptions import ExportNotSupported
@@ -365,12 +364,11 @@ class TestResources(unittest.TestCase):
         self.response.json.return_value = responses['project']['get']
         self.assertEqual(self.redmine.project.get(1).url, '{0}/projects/foo'.format(self.url))
 
+    @mock.patch('redmine.open', mock.mock_open(), create=True)
     def test_project_export(self):
-        self.response.content = 'foo'
-        self.response.text = b'foo'.decode()
         self.response.json.return_value = responses['project']['all']
-        self.assertEqual(self.redmine.project.all().export('txt', 'bytes'), 'foo')
-        self.assertEqual(self.redmine.project.all().export('txt', 'text'), b'foo'.decode())
+        self.response.iter_content = lambda chunk_size: (str(num) for num in range(0, 5))
+        self.assertEqual(self.redmine.project.all().export('txt', '/foo/bar'), '/foo/bar/projects.txt')
 
     def test_project_parent_converts_to_resource(self):
         from redmine.resources import Project
@@ -529,15 +527,13 @@ class TestResources(unittest.TestCase):
         self.response.json.return_value = responses['issue']['get']
         self.assertEqual(self.redmine.issue.get(1).url, '{0}/issues/1'.format(self.url))
 
+    @mock.patch('redmine.open', mock.mock_open(), create=True)
     def test_issue_export(self):
-        self.response.content = 'foo'
-        self.response.text = b'foo'.decode()
         self.response.json.return_value = responses['issue']['all']
-        self.assertEqual(self.redmine.issue.all().export('txt', 'bytes'), 'foo')
-        self.assertEqual(self.redmine.issue.all().export('txt', 'text'), b'foo'.decode())
+        self.response.iter_content = lambda chunk_size: (str(num) for num in range(0, 5))
+        self.assertEqual(self.redmine.issue.all().export('txt', '/foo/bar'), '/foo/bar/issues.txt')
         self.response.json.return_value = responses['issue']['get']
-        self.assertEqual(self.redmine.issue.get(1).export('txt', 'bytes'), 'foo')
-        self.assertEqual(self.redmine.issue.get(1).export('txt', 'text'), b'foo'.decode())
+        self.assertEqual(self.redmine.issue.get(1).export('txt', '/foo/bar'), '/foo/bar/1.txt')
 
     def test_issue_parent_converts_to_resource(self):
         from redmine.resources import Issue
@@ -618,12 +614,11 @@ class TestResources(unittest.TestCase):
         self.response.json.return_value = responses['time_entry']['get']
         self.assertEqual(self.redmine.time_entry.get(1).url, '{0}/time_entries/1'.format(self.url))
 
+    @mock.patch('redmine.open', mock.mock_open(), create=True)
     def test_time_entry_export(self):
-        self.response.content = 'foo'
-        self.response.text = b'foo'.decode()
         self.response.json.return_value = responses['time_entry']['all']
-        self.assertEqual(self.redmine.time_entry.all().export('txt', 'bytes'), 'foo')
-        self.assertEqual(self.redmine.time_entry.all().export('txt', 'text'), b'foo'.decode())
+        self.response.iter_content = lambda chunk_size: (str(num) for num in range(0, 5))
+        self.assertEqual(self.redmine.time_entry.all().export('txt', '/foo/bar'), '/foo/bar/time_entries.txt')
 
     def test_enumeration_version(self):
         self.assertEqual(self.redmine.enumeration.resource_class.redmine_version, '2.2')
@@ -770,12 +765,11 @@ class TestResources(unittest.TestCase):
             '{0}/projects/Foo/wiki/Foo'.format(self.url)
         )
 
+    @mock.patch('redmine.open', mock.mock_open(), create=True)
     def test_wiki_page_export(self):
-        self.response.content = 'foo'
-        self.response.text = b'foo'.decode()
         self.response.json.return_value = responses['wiki_page']['get']
-        self.assertEqual(self.redmine.wiki_page.get('Foo', project_id='Foo').export('txt', 'bytes'), 'foo')
-        self.assertEqual(self.redmine.wiki_page.get('Foo', project_id='Foo').export('txt', 'text'), b'foo'.decode())
+        self.response.iter_content = lambda chunk_size: (str(num) for num in range(0, 5))
+        self.assertEqual(self.redmine.wiki_page.get('Foo', project_id='Foo').export('txt', '/foo'), '/foo/Foo.txt')
 
     def test_wiki_page_parent_converts_to_resource(self):
         from redmine.resources import WikiPage
@@ -1178,12 +1172,11 @@ class TestResources(unittest.TestCase):
         self.response.json.return_value = responses['news']['filter']
         self.assertEqual(self.redmine.news.filter(project_id=1)[0].url, '{0}/news/1'.format(self.url))
 
+    @mock.patch('redmine.open', mock.mock_open(), create=True)
     def test_news_export(self):
-        self.response.content = 'foo'
-        self.response.text = b'foo'.decode()
         self.response.json.return_value = responses['news']['all']
-        self.assertEqual(self.redmine.news.all().export('txt', 'bytes'), 'foo')
-        self.assertEqual(self.redmine.news.all().export('txt', 'text'), b'foo'.decode())
+        self.response.iter_content = lambda chunk_size: (str(num) for num in range(0, 5))
+        self.assertEqual(self.redmine.news.all().export('txt', '/foo/bar'), '/foo/bar/news.txt')
 
     def test_news_str(self):
         self.response.json.return_value = responses['news']['filter']